requirements.txt
python-dotenv>=0.20.0,<0.21.0
requests>=2.27.1,<2.28.0
